Add search for home and discover, update UI elements, hide tab bar on detail pages

+++ EdufyApp/EdufyApp/Models/Academy/AcademyDetail.swift
@@ -16,7 +16,7 @@ struct AcademyDetail: Identifiable, Decodable {
     
     let totalApplications: Int
     let totalStudents: Int
-    let graduationRate: Int
+    let graduationRate: Double
 }
